更新，为不同物品应用不同的Tooltips

+++ src/main/java/com/xiao_xing/BetterTooltipBox/client/render/tooltipRender/Textrue/ITexture.java
@@ -6,9 +6,9 @@ public interface ITexture {
 
     ResourceLocation getResourceLocation();
 
-    int getWidth();
+    float getWidth();
 
-    int getHeight();
+    float getHeight();
 
     float getMaxU();
